+++ example/lib/page/home/logic.dart
@@ -1,5 +1,8 @@
+import 'dart:async';
+
 import 'package:get/get.dart';
 
 class HomeLogic extends GetxController {
-
+  final tick = 0.obs;
+  Timer? timer;
 }